requirements.txt
pyocd
pyserial
saleae
